fixed code-review recommendations

+++ posts/views.py
@@ -12,11 +12,14 @@ User = get_user_model()
 
 @cache_page(20)
 def index(request):
-    post_list = Post.objects.all()
+    post_list = Post.objects.select_related('group').all()
     paginator = Paginator(post_list, 10)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
-    return render(request, 'index.html', {'page': page, 'paginator': paginator, 'post': post_list})
+    return render(request, 'index.html', {
+        'page': page,
+        'paginator': paginator,
+        'post': post_list})
 
 
 def group_posts(request, slug):
@@ -34,14 +37,12 @@ def group_posts(request, slug):
 
 @login_required
 def new_post(request):
-    if request.method == 'POST':
-        form = PostForm(request.POST)
-        if form.is_valid():
-            user_post = form.save(commit=False)
-            user_post.author = request.user
-            user_post.save()
-            return redirect('index')
-    form = PostForm()
+    form = PostForm(request.POST or None, files=request.FILES or None)
+    if form.is_valid():
+        user_post = form.save(commit=False)
+        user_post.author = request.user
+        user_post.save()
+        return redirect('index')
     return render(request, 'new_post.html', {'is_edit': False, 'form': form})
 
 
@@ -52,12 +53,21 @@ def profile(request, username):
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     following = False
-    if request.user.is_authenticated and Follow.objects.filter(user=request.user, author=author).count() > 0:
+    if request.user.is_authenticated and Follow.objects.filter(
+            user=request.user,
+            author=author
+    ).exists():
         following = True
     return render(
         request,
         'profile.html',
-        {'page': page, 'post': post_list, 'author': author, 'paginator': paginator, 'following': following}
+        {
+            'page': page,
+            'post': post_list,
+            'author': author,
+            'paginator': paginator,
+            'following': following
+        }
     )
 
 
@@ -67,7 +77,15 @@ def post_view(request, username, post_id):
     post_list = post.author.posts.all()
     form = CommentForm()
     comment_list = post.comments.all()
-    return render(request, 'post.html', {'author': author, 'post': post, 'posts': post_list, 'comments': comment_list, 'form': form})
+    return render(
+        request,
+        'post.html',
+        {'author': author,
+         'post': post,
+         'posts': post_list,
+         'comments': comment_list,
+         'form': form}
+    )
 
 
 @login_required
@@ -75,27 +93,36 @@ def post_edit(request, username, post_id):
     post = get_object_or_404(Post, pk=post_id, author__username=username)
     if post.author != request.user:
         return redirect('post', username=username, post_id=post_id)
-    form = PostForm(request.POST or None, files=request.FILES or None, instance=post)
+    form = PostForm(
+        request.POST or None,
+        files=request.FILES or None,
+        instance=post)
     if form.is_valid():
         form.save()
         return redirect('post', username=username, post_id=post_id)
-    return render(request, 'new_post.html', {'is_edit': True, 'post': post, 'form': form})
+    return render(
+        request,
+        'new_post.html',
+        {'is_edit': True, 'post': post, 'form': form}
+    )
 
 
 @login_required
 def add_comment(request, username, post_id):
     post = get_object_or_404(Post, pk=post_id, author__username=username)
     comments = post.comments.all()
-    if request.method == 'POST':
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            new_comment = form.save(commit=False)
-            new_comment.post = post
-            new_comment.author = request.user
-            new_comment.save()
-            return redirect('post', username=username, post_id=post_id)
-    form = CommentForm()
-    return render(request, 'comments.html', {'form': form, 'post': post, 'comments': comments})
+    form = CommentForm(request.POST or None)
+    if form.is_valid():
+        new_comment = form.save(commit=False)
+        new_comment.post = post
+        new_comment.author = request.user
+        new_comment.save()
+        return redirect('post', username=username, post_id=post_id)
+    return render(
+        request,
+        'comments.html',
+        {'form': form, 'post': post, 'comments': comments}
+    )
 
 
 @login_required
@@ -104,14 +131,19 @@ def follow_index(request):
     paginator = Paginator(post, 10)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
-    return render(request, "follow.html", {'post': post, 'paginator': paginator, 'page': page})
+    return render(
+        request,
+        "follow.html",
+        {'post': post, 'paginator': paginator, 'page': page}
+    )
 
 
 @login_required
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
-    subscribe_count = Follow.objects.filter(author=author, user=request.user).count()
-    if request.user != author and subscribe_count == 0:
+    subscribe_exist = Follow.objects.filter(
+        author=author, user=request.user).exists()
+    if request.user != author and not subscribe_exist:
         follow_object = Follow.objects.create(author=author, user=request.user)
         author.following.add(follow_object)
     return redirect('profile', username=author.username)
@@ -120,7 +152,9 @@ def profile_follow(request, username):
 @login_required
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
-    if request.user != author:
+    is_following = Follow.objects.filter(
+        author=author, user=request.user).exists()
+    if request.user != author and is_following:
         Follow.objects.filter(author=author, user=request.user).delete()
     return redirect('profile', username=author.username)
